Moved MimeTypes from API to Shared module so that I can start being more intelligent with file uploads.
