+++ API/src/main/kotlin/org/liamjd/cantilever/routing/APIGatewayProxyEventExtensions.kt
@@ -1,6 +1,7 @@
 package org.liamjd.cantilever.routing
 
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent
+import org.liamjd.cantilever.common.MimeType
 
 data class Header(val name: String, val value: String)
