+++ API/src/main/kotlin/org/liamjd/cantilever/routing/RequestPredicate.kt
@@ -1,6 +1,7 @@
 package org.liamjd.cantilever.routing
 
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent
+import org.liamjd.cantilever.common.MimeType
 import kotlin.reflect.KType
 
 /**
