+++ API/src/main/kotlin/org/liamjd/cantilever/routing/simple/SimpleRouter.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.MissingFieldException
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.serializer
-import org.liamjd.cantilever.routing.MimeType
+import org.liamjd.cantilever.common.MimeType
 import kotlin.reflect.typeOf
 
 /**
